Enforce User Password Expiry [#326 responsible:none state:fixed milestone:v2.0]
Added password expiry functionality. The system.appuser config in the SolaRealm
of Glassfish should be changed to reference the system.active_users view. This
view implements functionality to support password expiry from the JDBCRealm.

The length of time a users password remains valid is controlled by the
pword-expiry-days system.setting. Default is 90 days. If this setting is
NULL or disabled, no password expiry is used.

Users will be warned if their password is within 14 days of expiry and
prompted to renew it. For most users, once their password expires, they
will no longer be able to login to SOLA. Users with the ManageSecurity
and/or NoPasswordExpiry role will still be able to login to SOLA once
thier password is expired. (i.e. Admin users and user accounts required
for external systems integration with SOLA web services.)

Users will also be prompted to change their password if they were not
the last person to set it (i.e. to cover situations where an Admin
resets a users password).